examples/tools/string-tools/pyproject.toml
[project]
name = "string-tools"
version = "0.1.0"
description = "Simple string manipulation tools for MCP Fabric agents"
requires-python = ">=3.12"
dependencies = [
    "strands-agents",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
